requirements.txt
edge-tts
gTTS
fastapi
uvicorn
openai 
groq
faster-whisper
python-dotenv
python-multipart
python-jose[cryptography]
requests
supabase
pydantic-settings
